Add files via upload
These are the original python scripts I used to create the HTML to display photos or proofs on my HubSpot hosted photography site. These scripts take a name and directory to complete the page. Run these scripts from a terminal window in the directory which contains the photos you want added to the site.